#1 Remove dependencies of example app from common.
